requirements.txt
pytest
numpy

# Project Logbook
pandas
tqdm
docx2pdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Documentation
sphinx
sphinxcontrib-napoleon
